requirements.txt
Flask
delegator.py
pexpect  # required by delegator.py
